requirements.txt
web3
python-dotenv
py-solc-x
pandas
scikit-learn
Flask
